Add menu filters to other menu pages, session authentication, code cleanup.

+++ HTML/js/env.js
@@ -1,13 +1,16 @@
 document.addEventListener('DOMContentLoaded', defEnv);
 
 function defEnv () {
-    var logout = document.getElementById("logout");
-    logout.addEventListener('click', function(event) {
-        sessionStorage.clear();        
-    });
-    
-    var user = sessionStorage.getItem('user');
-    if (user) {
-        document.getElementById('username').innerHTML = user
+    if (sessionStorage.length == 0) {
+        window.location.href = "login.html";
+    } else {
+        var logout = document.getElementById("logout");
+        logout.addEventListener('click', function(event) {
+            sessionStorage.clear();        
+        });
+        var user = sessionStorage.getItem('user');
+        if (user) {
+            document.getElementById('username').innerHTML = user
+        }
     }
 }
